Refactored test jobs that run inside the cluster to use a different build tag
The test jobs have been placed into another package and they now use a
`job` tag. This way there is no longer a need to call `SkipTestJobIfMustBeSkipped`
as those tests won't be executed as an integration test.

Resolves #424
