+++ Makefile
@@ -10,10 +10,10 @@ all: build-cmd build-get build-controllers build-tests
 
 build-tests:
 	mkdir -p ${TEST_BINARIES_FOLDER}
-	go test -c -tags=integration -v ./test/integration/tcp_echo -o ${TEST_BINARIES_FOLDER}/tcp_echo_test
-	go test -c -tags=integration -v ./test/integration/http -o ${TEST_BINARIES_FOLDER}/http_test
-	go test -c -tags=integration -v ./test/integration/bookinfo -o ${TEST_BINARIES_FOLDER}/bookinfo_test
-	go test -c -tags=integration -v ./test/integration/mongodb -o ${TEST_BINARIES_FOLDER}/mongo_test
+	go test -c -tags=job -v ./test/integration/tcp_echo/job -o ${TEST_BINARIES_FOLDER}/tcp_echo_test
+	go test -c -tags=job -v ./test/integration/http/job -o ${TEST_BINARIES_FOLDER}/http_test
+	go test -c -tags=job -v ./test/integration/bookinfo/job -o ${TEST_BINARIES_FOLDER}/bookinfo_test
+	go test -c -tags=job -v ./test/integration/mongodb/job -o ${TEST_BINARIES_FOLDER}/mongo_test
 
 build-cmd:
 	go build -ldflags="${LDFLAGS}"  -o skupper ./cmd/skupper
